# src/openapi.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Listar usuarios (paginado)",
        "security": [
          {
            "basicAuth": [],
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "email | username | phone"
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "email",
                "username"
              ]
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Crear usuario",
        "description": "Solo administradores.",
        "security": [
          {
            "basicAuth": [],
            "accessCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "409": {
            "description": "Usuario ya existe"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtener un usuario",
        "security": [
          {
            "basicAuth": [],
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUser"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Actualizar usuario",
        "description": "El propio usuario puede editar su perfil. Campos sensibles (email/password) tienen endpoints dedicados. Cambios de role/status requieren admin.",
        "security": [
          {
            "basicAuth": [],
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUser"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado para cambiar role/status"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "409": {
            "description": "Valor duplicado (username, etc.)"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Eliminar (soft-delete) usuario",
        "description": "Deshabilita la cuenta, revoca sesiones y (opcional) anonimiza PII.",
        "security": [
          {
            "basicAuth": [],
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario deshabilitado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMessage"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "isOnline": {
            "type": "boolean"
          },
          "role": {
            "type": "string"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "profileImage": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ApiResponseMessage": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "OK"
          }
        }
      },
      "ApiResponseUser": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Usuario creado"
          },
          "data": {
            "$ref": "#/components/schemas/UserPublic"
          }
        }
      },
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "perPage": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "UsersListResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/PaginationInfo"
              }
            }
          }
        }
      },
      "UserCreateRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "seller",
              "buyer",
              "support"
            ],
            "default": "buyer"
          },
          "username": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "description": "Campos editables por el usuario. Cambios de role/status requieren admin.",
        "properties": {
          "username": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "profileImage": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "seller",
              "buyer",
              "support"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    }
  },
...
  "info": {
    "title": "commerceHub API",
    "version": "1.0.0",
    "description": "Documento base. Los módulos agregan paths; los schemas viven aquí."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ]
}